Make View use float internally
I don't know if this is right decision or not...

+++ src/Gfx/Gfx.cs
@@ -14,8 +14,8 @@ namespace Flora.Gfx {
 
         // view
         internal static View activeView = null;
-        internal static int activeViewOffsetX { get { return activeView == null ? 0 : activeView.centerX - activeView.offsetX; } }
-        internal static int activeViewOffsetY { get { return activeView == null ? 0 : activeView.centerY - activeView.offsetY; } }
+        internal static int activeViewOffsetX { get { return activeView == null ? 0 : (int)(activeView.centerX - activeView.offsetX); } }
+        internal static int activeViewOffsetY { get { return activeView == null ? 0 : (int)(activeView.centerY - activeView.offsetY); } }
 
         /// <summary>
         /// Flip options for drawing textures.
